+++ RELEASE_NOTES.md
@@ -48,12 +48,12 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `EmojiKeyboardStyle` has new context-based standard functions.
 * `ExternalKeyboardContext` is now available on macOS and tvOS as well.
 * `KeyboardAction` gesture actions are now available on all platforms.  
-* `KeyboardAction` gesture actions now use a `KeyboardActionTrigger` instead of a `KeyboardInputViewController`.  
-* `KeyboardActionTrigger` is a new protocol that lets us decouple actions from the input view controller.
+* `KeyboardAction` gesture actions now use a `KeyboardController` instead of a `KeyboardInputViewController`.  
 * `KeyboardAppearance` will now apply to `.nextKeyboard` as well. 
 * `KeyboardCalloutContext` is a new context that lets lets us pass around a single context for input and action callouts.
+* `KeyboardController` is a new protocol that lets us decouple actions from the input view controller.
 * `KeyboardInputViewController` has a new `calloutContext`.  
-* `KeyboardInputViewController` implements `KeyboardActionTrigger` which gives it a bunch of new functions.
+* `KeyboardInputViewController` implements `KeyboardController` which gives it a bunch of new functions.
 * `NextKeyboardButton` now takes any custom content and only applies the behavior on iOS and tvOS.
 * `GestureButton` is now available on watchOS 7.
 * `InterfaceOrientation` has a new `landscape` case.